+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Volumes/Evo512/sdk/flutter/flutter_beta"
-export "FLUTTER_APPLICATION_PATH=/Users/cai/Documents/GitHub/fluttercandies/ff_annotation_route/example"
-export "FLUTTER_TARGET=lib/main.dart"
+export "FLUTTER_ROOT=E:\Flutter\flutter_source\master"
+export "FLUTTER_APPLICATION_PATH=E:\Flutter\FlutterCandies\ff_annotation_route\example"
+export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "SYMROOT=${SOURCE_ROOT}/../build/ios"
-export "FLUTTER_FRAMEWORK_DIR=/Volumes/Evo512/sdk/flutter/flutter_beta/bin/cache/artifacts/engine/ios"
+export "SYMROOT=${SOURCE_ROOT}/../build\ios"
+export "FLUTTER_FRAMEWORK_DIR=E:\Flutter\flutter_source\master\bin\cache\artifacts\engine\ios"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"