+++ lib/src/command/help.dart
@@ -1,0 +1,32 @@
+import 'package:ff_annotation_route/src/command/command.dart';
+
+class Help extends Command {
+  @override
+  String get description => 'Print this usage information.';
+
+  @override
+  String get full => '--help';
+
+  @override
+  String get short => '-h';
+}
+
+String get help {
+  final maxLengthCommand = commands.reduce((value, element) =>
+      value.command.length > element.command.length ? value : element);
+
+  final sb = StringBuffer();
+  for (var i = 0; i < commands.length; i++) {
+    final command = commands[i];
+    sb.write(
+        '${command.command}${' ' * (maxLengthCommand.command.length - command.command.length + 5)}${command.description}\n');
+  }
+
+  return '''\nManage your Flutter app development with ff_annotation_route.
+
+Usage: ff_route <command> [arguments]
+
+Available commands:
+${sb.toString()}
+''';
+}